feat: Add LoginRedirect component and update main and home pages
