+++ src/pages/Myspace.tsx
@@ -1,10 +1,15 @@
 import Toolbar from '@/components/Myspace/Toolbar';
+import { useSelector } from 'react-redux';
+import { selectAllUsers } from '@components/users/userSlice';
 
 const Myspace = () => {
+  const users = useSelector(selectAllUsers);
+
+  console.log(users);
   return (
-    <div>
+    <>
       <Toolbar />
-    </div>
+    </>
   );
 };
